+++ app/lib/ui/components/buttons/component_filled_tonal_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter/scheduler.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:ods_flutter/components/button/ods_filled_tonal_button.dart';
-import 'package:ods_flutter_demo/ui/components/buttons/component_button_customization.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/button_customization.dart';
 import 'package:ods_flutter_demo/ui/components/utilities/customization_bottom_sheet.dart';
 import 'package:ods_flutter_demo/ui/main_app_bar.dart';
